Coins screen

+++ app/src/main/java/com/mathroda/dashcoin/data/remote/dto/ChartDto.kt
@@ -3,7 +3,7 @@ package com.mathroda.dashcoin.data.remote.dto
 import com.mathroda.dashcoin.domain.model.Charts
 
 data class ChartDto(
-    val chart: List<List<Double>>
+    val chart: List<List<Float>>,
 )
 
 fun ChartDto.toChart(): Charts{
